Make storybook flat (refactor sub-folders)
