+++ src/stories/actor.stories.ts
@@ -1,10 +1,10 @@
 import {Component} from '@angular/core';
 import {moduleMetadata} from '@storybook/angular';
-import {AtftDataCenterActorModule} from '../../../projects/atft/src/lib/actor/data-center';
-import {AnimationService} from '../../../projects/atft/src/lib/animation/animation.service';
+import {AtftDataCenterActorModule} from '../../projects/atft/src/lib/actor/data-center';
+import {AnimationService} from '../../projects/atft/src/lib/animation/animation.service';
 // NOTE: Do direct import instead of library (allows to watch component and easy to develop)
-import {AtftModule} from '../../../projects/atft/src/lib/atft.module';
-import {worldSceneWrapper} from '../scene-wrapper/world-scene-wrapper';
+import {AtftModule} from '../../projects/atft/src/lib/atft.module';
+import {worldSceneWrapper} from './scene-wrapper/world-scene-wrapper';
 
 
 @Component({
