+++ src/stories/object.stories.mdx
@@ -1,5 +1,4 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
-import {BoxMeshComponent} from "../../../projects/atft/src/lib/object/mesh";
 
 <Meta title="Object/Abstract"/>
